move wapo.pyramid.errors.JSONError to wapo.pyramid.views.JSONErrorView
includes a replacement import for backwards compatibility
